Use regular expressions for multiple drivers in one Printer Application
Retro-fitting Printer Applications can easily contain more than one
printer driver and so support certain printers with more than one
driver. This does not only happen with Ghostscript (which contains a
large amount of built-in printer drivers) but also with the fact that
we have several tiny, unmaintained printer driver projects which we
want to put together and retro-fit in one single Printer Application
to not clutter the Snap Store and to ease tyhe maintenance of the
Snaps.

Before, we extracted only manufacturer and model of the PPDs to make
up the "driver" entries. This worked well for PostScript and also for
single-driver Printer Applications, but if we have several classic
drivers in our Printer Application and they all support a certain
printer, only one arbitrary PPD gets considered and the others skipped
as duplicates, not giving access to all drivers for this printer, and
the arbitrary driver which gets available is not necessarily the best
one.

This commit does two things: It makes all suitable drivers for each
printer listed in a clean, user-friendy, uncluttered way and it
prioritizes drivers against others so that the best driver gets
selected when auto-adding printers or use the automatic driver
selection on the "Add Printer" page of teh web interface.

This is implemented by the use of regular expressions. In the
configuration data structure for our Printer Application we define one
regular expression for making up the make/model/driver listing entries
(and also the PAPPL-internal driver names), with matches the whole
part after the model name in the PPD's Nickname and so determines the
separation of model name and driver info and contains a pair of
parentheses to mark the part of the driver info string which should be
shown in the listing (and used for the PAPPL-internal driver name).

This way we get clean, user-friendly list entries, without repeatedly
appearing words which the user does not know anything of (like
"Foomatic") or version numbers (like in Gutenprint PPD files).

For the priorization on auto-selection of drivers we allow adding a
list of regular expressions to the configuration data structure where
for each PPD matching make and model of the printer each regular
expression in the list is matched against the PAPPL-internal driver
name of the PPD file. The earlier in the list the matching regular
expression appears the higher the priority is for this PPD file.

In addition, two new control points for the configuration of the
Printer Application got added: We can now suppress the
auto-determination of a "generic" PPD file for the "generic" driver in
the beginning of the list and we also can opt for skipping PPD files
on which the regular expression for the make/model/driver listings
does not match.