Do not add "USER-ADDED" in driver list if we do not support user-added PPDs
Before checking if a PPD is in the directory for user-added PPDs we
check at first whether we support user-adding of PPDs at all, via the
PR_COPTIONS_WEB_ADD_PPDS flag. As otherwise we would check against an
empty user PPD directory name which always matches. We also check the
end of the user PPD directory name in the PPD record path.