Fixed crash caused by PPD on a queue being changed and memory leak
When a print queue is created in the Printer Application with a
certain PPD file and the PPD files is later on removed from the
package yhen on the next start the Printer Application falls back to
driver "auto", auto-selecting a new PPD file for the queue.

If the new PPD file has different sets of page sizes and types or
different page margins than the original ones, media-ready entries
with media not valid in the new PPD file can be loaded from the state
file and can mess up pointers, leading to a crash by freeing an
invalid pointer when the printer data structure is freed, either when
removing the print queue or when terminating the Printer Application.

This is now fixed by verifying and correcting all media-ready entries
when updating the driver data structure after loading the sate file.

In addition, fixed a memory leak:

For the names of media size variants (with '.' in the PPD name) memory
were allocated which got lost when the variant was skipped after
checking its margins and processing the following media size in the
PPD. Now we allocate the memory for the media size name only after
checking the margins and skipping the variant, if it is one.

Also fixed a typo in a comment and silenced a (harmless) warning.