Fixed setting up media-ready entries after loading the state file
As of https://github.com/michaelrsweet/pappl/issues/164 and the final
fix https://github.com/michaelrsweet/pappl/commit/52b7492bc the ready
media in the currently selected tray (the source in default_media)
should be also the default media. On any change in the ready media,
the entry of the default source is copied into the default_media by
the papplPrinterSetReadyMedia() function.

To not break the setting when updating the list of ready media by
running the pre_driver_setup() function in update mode, we copy the
media_ready entry of the default source into the default_media.