On printer model/driver list add comma only if actually needed
If the regular expression for displaying the driver name has no
parentheses, the whole string is displayed as driver name. In this
case (and also if the part matching the parentheses includes the first
characters right from the end of the model name on) we do not want to
display an added comma between model name and driver name, as we are
preserving the original separator chracters (and displaying both looks
awkward).