Remove memmove() calls where Valgrind tells of memcpy() on overlapping region
In the pr_driver_setup() function when looking for IPP names for
vendor-specific PPD options we get Valgrind messages about memcpy()
use with overlapping memory reasons though we are using memmove()
which is made for overlapping memory regions. Probably a bug somewhere
in the libraries or in Valgrind.

Replaced the memmove() calls by simple loops moving the characters.