Fix crash when queue's PPD file data is freed from memory
When shutting down or when removing a printer the PPD data structures
(main data structure and PPD cache of the queue are freed from memory.

This is done by the ppdClose() function which exists equally named in
libcups and libppd. The ppdClose() function removes the main structure
and calls a function to remove the cache.

Problem here is that the linker can link the ppdClose() function from
libppd or from libcups, and we need the onr from libppd as it calls
the cache functions from libppd due to the fact that in libppd the PPD
cache data structure got modified, it got a new field for the content
optimize presets. If one tries to free the new data structure with the
function of libcups, one gets a crash.

To work around this we make use of the fact that in libppd the PPD
cache functions are promoted from internal functions to API functions
and therefore renamed from "_ppd..." to "ppd...". So we can safely
call these functions from libppd and therefore we call
ppdCacheDestroy() on the cache before calling ppdClose() to make sure
that the cache gets freed by libppd and not by libcups.