Several improvements on building driver data record from the PPD
When testing Gutenprint PPD files discovered several issues:

1. "Borderless" switch in web interface

If all page sizes in the PPDs have only margins, we do not need a
switch for borderless or non-borderless in the web interface

2. Limits for Raster input resolution

The resolutions registered in the driver data record are actually only
used for Apple/PWG Raster and image input data. As in case of
Apple/PWG Raster jobs the client needs to provide the data in the
given resolution, a client can get easily overloaded by having
extremely high resolutions (as defined in the PPD file)
here. Therefore we limit these resolutions to a maximum of 1440 dpi
for high, 720 dpi for normal, and 360 dpi for draft quality.

The resolutions in the PPD files are the hardware resolutions of the
printers, which are used for high-quality dithering. For the input
data usually resolutions of 1440 dpi for drawings/lines and 720 dpi
for photos/images are good enough.

Higher resolutions we reduce by halving them until they are below the
limit. This way the input resolution and the device resolution stay
multiples of 2 and so the pwgtoraster() filter function of
cups-filters is able to convert the resolution if needed.

3. Duplicate media property entries

We check for media size, source, and type whether the PWG names for
any 2 choices in the PPD option are duplicates and do not add such
duplicate entries to the lists in the driver data record. In the web
interface they only confuse the user and do not give access to
different settings.

In addition, we log the addition of all media property items and fix
some typos in the comments.