Explicitly shut down CUPS backend on end of job to avoid crash
When a CUPS backend is used, for back and side channel support and for
logging the filter_data structure is shared betwwen the filter
functions and the backend. In normal PAPPL operation filtering ends
and the related dynamic memory is freed before the device (to which
the CUPS backend belongs) is closed. As the CUPS backend uses the
filter_data structure of the filtering (at least for logging) we need
to shut down the backend early when shutting down the filtering,
before freeing the filter_data structure, otherwise we will get a
crash, accessing to already freed memory.