Allow setting callback functions for extra system and printer setup steps
Now two more callback functions can be set in the configuration of the
Printer Application, one executed during system setup, to do extra
steps here, especially adding system-wide (not print-queue-specific)
buttons and/or pages to the web interface, and the other executed
during printer setup to do extra steps there, like
print-queue-specific web interface extensions.

Examples use cases are for example the download of the proprietary
plugin for HPLIP as a system-wide extension or head cleaning/alignment
as a print-queue-specific extension.

Note that this is an API change which requires changes on all Printer
Applications using this library.