When generating driver list, never use sinple "*Product" entry in PPD
When generating the driver (model) list from the PPD collections, we
loop through the "*Product" lines of the PPds to find whether there
are other model names supported by the same PPd file, creating more
than one driver list entry for the same PPD file.

In this loop we already catch cases of a single "*Product" line in the
PPD file and drop this "*Product" line then, as it is either the model
we already have from the "*NickName" or something weird.

But, in contrary to this, we grab the first "*Product" line instead of the
"*NickName" already before actually looping through the "*Product" lines,
meaning that we actually catch a weird "*Product" entry.

This commit fixes this by also checking whether the first "*Product"
line is the only one before we grab it then.