Let pr_system_cb() set the CUPS_SERVERBIN env variable
If CUPS_SERVERBIN is not set and FILTER_DIR is set to a path ending
with "/filter", we set CUP_SERVERBIN to the value of FILTER_DIR minus
the "/filter" at the end. This way CUPS filters in the FILTER_DIR,
called by filterExternalCUPS() of cups-filters get an environment as
close as to the one of CUPS as possible.