Use device ID as display string only if manufacturer matches
Some printers sold by a certain manufacturer (and having that name
visible on the device) are actually made by another manufacturer (and
having that name in their device ID). Using the device ID's
manufacturer/model as display string in the PPD list makes the printer
appear with a name under which it is not know, and the known name of
the printer does not appear in the list.

In this case automatic assignment of the PPD (set "Driver" to "Auto")
works, but if it does not work for another reason (ex. backend does
not report actual device ID), the user does not find his printer in
the list for manual assignment.

Therefore we check the manufacturer name in the device ID whether it
matches the one of the NickName (after normalizing both) and only with
a match we accept the device ID's data for the display string.

Example for such a mismatch are the PPD files of the driver
rastertosag-gdi, for the printers Ricoh Aficio SP 1000S and SP 1100S
with "MFPrinter Laser Pro LL(2)" as device IDs. It seems that Ricoh
has bought these printers from a white-label manufacturer. Especially
they are also of the very few Ricoh printers which do not understand
PostScript nor PCL.