Use OUTPUT_FORMAT_PDF_IMAGE with ghostscript() when streaming
Call the ghostscript() filter function with the
OUTPUT_FORMAT_PDF_IMAGE parameter (instead of just OUTPUT_FORMAT_PDF)
when printing PDF in streaming mode. This makes the PDF output being
streamed to the printer (or to the driver), so that if the printer
supports streaming PDF it prints the pages right away instead of
waiting until the full job arrives. To raise the probability that the
printer actually streams the PDF job the ghostscript() filter function
sends PCLm if the printer supports it (according to the PPD file or
IPP attributes).

This requires Ghostscript 9.55.0 or later.