Added dynamic PPD generator for *.drv files
Many classic CUPS drivers provide their PPD files as driver
information files (*.drv). These files are not supported by libppd and
so also not by Printer Applications using libpappl-retrofit.

To use such drivers without needing to pre-build the PPD files, for
example if one simply lets a Printer Application use the CUPS drivers
installed for the local CUPS, as the test-printer-app included with
pappl-retrofit does in its standard installation, one can use this
program.

It is simply a dynamic PPD generator (what CUPS drivers put into
/usr/lib/cups/driver) which goes through /usr/share/cups/drv and lists
all PPD files made available by the *.drv files ("list" argument) and
builds them on-demand ("cat" argument with URI of the desired PPD
file). So makiung it available to the Printer Application allows the
Printer Application to use the PPDs described by the *.drv files.

It is not yet integrated in the build system of pappl-retrofit. For
now you can compile it with

    g++ -o drv drv.cxx $CUPS_SOURCE/ppdc/libcupsppdc.a \
        -DCUPS_DATADIR='"/usr/share/cups"' -I $CUPS_SOURCE -lcups -lppd

and install it into /usr/share/ppd. Run the test-printer-app then and
it shows also the printers/drivers from the *.drv files when adding a
printer.

Do not install it into /usr/lib/cups/driver, to avoid duplicate
listings of printers by CUPS.

Building this program needs the source code of CUPS 2.4.x or
older. Needs cups-filters 2.x or newer.