Added a bannertopdf() support and a dynamic PDF test page
To make it easier to create test page graphics content/customize the
test page and to not require a PostScript interpreter in the Printer
Application only to be able to print a dynamic (with job and printer
information and in job's page size) test page, we have added a
PDF-based alternative.

With this a Printer Application which supports non-PostScript printers
can use a pure PDF interpreter (like Poppler or MuPDF) instead of
Ghostscript for PDF input without loss of the ability to print a
dynamic test page.

In addition, nowadays it is much easier to export a document created
with a graphics editing application in PDF than in PostScript.

Therefore we now support PDF test pages using bannertopdf() of
cups-filters. bannertopdf() uses a one-page PDF file as template,
scales it to the job's/the printer's page size, and then adds job- and
printer-specific information in text form onto the lower half of the
page. On jobs with N-up ("number-up") and/or duplex ("sides") options
set, the test page is replicated as many times as needed to fill a
full sheet to be able to test the effect of these options.

The test page needs to be a one-page PDF, ideally vector graphics
(scalable) in A4 or Letter (so that in most cases no extreme scaling
is needed). The graphics should be concentrated on the upper half and
the lower half should be white or light-toned and not busy, so that
added black text is easily readable. A frame all around the page is
also possible, the text does not get too close border. The graphics
should be suitable for testing the color output, photo output, and
resolution of the printer, so it depends on the range of printers
supported by the Printer Application what to use.

To the PDF file some lines have to be added at the end, to mark it as
a banner or test page, so that bannertopdf() gets triggered to add the
dynamic text content and not pass the page through like a regular PDF
job. Also lines to select the desired dynamic content can be added.

It must be always added:

    %%#PDF-BANNER

To get all dynamic items shown, add:

    %%Show printer-name printer-info printer-location printer-make-and-model printer-driver-name printer-driver-version paper-size imageable-area job-id options time-at-creation time-at-processing job-originating-user-name job-name job-originating-host-name job-billing job-uuid

Remove items from this line to select what actually should get added.

The lines are PDF comments and so do not affect the subsequent
filters. Taking away the PDF comment mark ("%%") the lines have the
same syntax as PDF banner instruction files of cups-filters
(application/vnd.cups-pdf-banner).

The file data/testpage.pdf is an example for such a PDF test page file.

Naturally the Printer Application has to be configured to accept PDF
as input data format.

The Test Printer Application uses testprint.pdf as test page now.

The pr_create_job_data() function now sets the environment variables
PRINTER and PRINTER_LOCATION as CUPS would set them, for use by the
filters. This especially allows bannertopdf() to show the printer
location text as set for the printer in the Printer Application.

Also the open() call of the (commented out) debug dropping of the
print job content into /tmp/printout file in
pr_print_filter_function() got updated now to work with current
compilers.