In debug logging mode save copies of the jobs sent to printer
If we run the Printer Application in debug logging mode ("-o
log-level=debug" or switching on the logging page of the web
interface) from every job a copy of the data actually sent to the
printer gets saved in a file (debug-copy-PRINTER-JOB.prn) in the spool
directory. These files are kept for 24 hours (clean-up done with every
new job and every start of the Printer Application, independent of log
level).