Generate artificial device ID for PPDs without explicit device ID
If a PPD has no explicit device ID, the device ID field of the driver
list stays empty and the autoadd callback matches manufacturer and
mode against the display name, which is the PPD's NickName in many
cases.

As with a display name of "HP LaserJet 2 (en)" one cannot distinguish
whether this is for the HP LaserJet 2 or Whether it is driver version
2 for the HPL LaserJet, a user's HP LaserJet reporting
"MFG:HP;MDL:LaserJet;" would also match this one and we get it set up
with the driver for the HP LaserJet 2.

To avoid this, we create artifical device IDs for the PPD files which
do not explicitly tell the printer's device ID. Here we fill in the
PPD's manufacturer name entry for "MFG:" and try to fill a clean model
name without driver specification for "MDL: ". We try at first the
first Product entry if it matches the beginning of the NickName, after
that we use the ModelName, and only if this fails, the NickName.