Improved matching of device ID against driver name
Before, the device ID of the HP LaserJet, "MFG:HP;MDL:LaserJet;",
matched the driver name of the HP LaserJet 2,
"hp--laserjet-2--en". Now we check for the two '-' characters after
the model name in the driver names, to prevent from this mis-match and
avoid false positives when automatically finding the correct driver
for a given printer (or whether a given Printer Application supports a
given printer).