Extract driver info from PPD's NickNames
Before we could reliably extract driver info (and separate it from the
model name) from the PPD's NickNames only by supplying a regular
expression (one per Printer Application) which makes the task complex
for Printer Applications containing many different drivers
(Ghostscript Printer Application, or a Printer Application making
available all drivers installed for CUPS on the system, as the Test
Printer Application).

Now, after the regular expression not having matched (or none
supplied) we check the NickName for ',', '(', and " - " and consider
this is separator between model name and driver info. And if such
characters are not in the NickName, but we have the ModelName from the
PPD, we consider as driver info what the NickName has more than the
ModelName.

We also add a ',' as separator if the driver info string found by the
regular expression has no separator.

With this, the driver name has double dashes ("--") always between
manufacturer and model and between model and driver info now.

And on every auto-match of a device ID with the pr_best_matching_ppd()
function we log the device ID and also the normalized make and mode
names to match against the driver name.