When setting up the driver verify and correct default resolution
When setting up a printer driver and loading the default settings from
the previous session the loaded default setting for the resolution was
not checked whether it is one of the supported resolutions.

Usually, there cannot be a wrong resolution default in the state file,
as on initial setup of the print queue the available resolutions are
taken from the PPD file and the default, too, and the web interface
only allows choosing one of the available resolutions as default.

Problem is if a driver gets removed from the Printer Application, or
renamed, and then after an update to this new version, a queue which
used this driver gets auto-assigned another driver and that new driver
does not support the default resolution of the queue any more. In this
case PAPPL's verification of the default settings failed and the setup
of the driver stayed incomplete, leading to a crash/core dump when
cleaning up on shutdown (or removal of the queue).

Now the default resolution gets verified (and corrected to the
resolution for "normal" print quality) during driver setup, before
passing the data structure to PAPPL.

The defaults for all the other attributes were already verified and
corrected before.