test-printer-app: Do not run driverless utility
The test-printer-app is supposed to make all installed CUPS drivers
available in a Printer Application, simply by installing it. For this
it browses the directories where PPD files and PPD file generators are
expected by CUPS.

Only in many systems there is also the driverless utility of
cups-filters installed and linked to /usr/lib/cups/driver, producing
PPD file listings for driverless IPP printers, and these we do not
want in a Printer Application. Therefore we set the NO_DRIVERLESS_PPDS
environment variable to tell the driverless utility to not do
anything, neither browse the network and let the Printer Application
wait for that nor list any PPD files.