Renamed test-printer-app into legacy-printer-app
As the included Printer Application does not only serve as an example
for how to use the library ot for testing, but instead, also makes all
classically installed printer drivers available in a Printer
Application, so that one can easily switch to the CUPS Snap from a
classically installed CUPS. Therefore we have renamed it, as it it
makes legacy CUPS drivers available to the CUPS Snap (and also to CUPS
3.x or newer).

For making classically installed drivers available the best is also to
use the classically installed CUPS backends. Therefore the Legacy
Printer Application only uses the CUPS backends. As it also continues
to be useful for development and testing, the
"--enable-pappl-backends-for-legacy-printer-app" option for
"./configure" was now added which makes the Legacy Printer Application
being built using both CUPS backends and PAPPL's built-in backends
(schemes).