Simplify filter chain build for streaming job processing
If we process a job in streaming mode (PWG/Apple Raster, images) we do
not simply throw the needed filters into a filterChain() call (as we
do for spooling mode, for PDF or PostScript), but check first whether
we have only the backend to call and no pre-filter before, which gives
us a chain of a single filter function, and for this case we do not
use filterChain().

First, when I have introduced the copying of the job data into a file
in debug logging mode (commit 4fb435163) I have forgotten to set up
the new parameter data structure for the backend filter function
pr_print_filter_function(), which lead to Issue #1 being reported, a
crash on a Raster job sent to a PostScript printer. Second, the
distinction is not needed, as it is no problem to call filterChain()
with a single filter.

So with this commit we simplify and call filterChain() regardless of
whether there are pre-filters or not. This simplifies the code and
should solve the issue.

The alternative approach of keeping the distinction but only creating
the parameter data structure independently would lead us also into
extra complication when having to free the memory of it after the job.