Added SNMP-based supply-level checking
We make use of the new feature of PAPPL 1.2 to do SNMP-based
supply-level queries on the printer.

The implementation in pappl-retrofit follows the implementation in the
hp-printer-app (https://github.com/michaelrsweet/hp-printer-app/),
file hp-printer-app.c.

The querying only works with network-connected printers and only with
backends of PAPPL itself (schemes: "snmp:", "dnssd:", "socket:"). It
does not work with embedded CUPS backends (scheme: "cups:") yet. It
also does not work on non-network printers (scheme: "usb:"). Therefore
we activate the feature only for appropriate printers, and only those
will therefore get a "Supplies" button in the web admin interface.

A click on the "Supplies" button shows the supply levels of the
printer as a chart. It also shows the manufacturer's part numbers of
the supplies as the printer provides them.

This should work with all printers where supply checking with CUPS and
classic printer drivers works (you can display the supply info in the
"Properties" dialog of system-config-printer then).