Create copy of the physical PPD when CUPS backend is used
Not only when the printer's PPD defines a CUPS filter to be used
(usually on non-PostScript printers a CUPS driver) but also if we use
a CUPS backend, we create a temporary copy of the physical PPD file so
that the backend can read it.

This is especially needed for the network CUPS backends (like
"socket") to enable supply level and status checking via SNMP. They
refuse to do it if there is no PPD or if the PPD contains
"*cupsSNMPSupplies: False".

This is a first step for supply level (and perhaps also enhanced
status) checking support in CUPS-driver-retro-fitting Printer
Applications.