Correct where status update is called.
In one of the previous commits a status updating function was
introduced and called before and after the execution of each job, to
update supply level info so that the supply levels can be displayed in
the web interface.

Unfortunately, the calls were misplaced (but only affecting queues
using CUPS backends), the before-job call was too early, before
connecting the filter data record to the CUPS backend, and the
after-job call was to late, after shutting down the CUPS backend and
disconnecting the filter data record from it.

This caused unexpected behavior up to crashes.

This should solve
https://github.com/OpenPrinting/ps-printer-app/issues/14