pr_print_filter_function(): Do not shut down CUPS backend at end of input data
As we now do check of device status and ink/toner levels before and
after each job we cannot already close the CUPS backend by when the
pr_print_filter_function() detects that the complete job data has been
passed on. So we remove this functionality from
pr_print_filter_function().

We shut down the CUPS backend anyway when the job ends, after the call
to update the device status.