Adapted to newest changes in cups-filters 2.x, fixed crasher
The filter data record which is supplied to all filters in a chain now
hold input data format (content_type) and output data format
(final_content_type). Therefore most filter functions will not take
the output data type as filter-instance-specific parameter any more.

- In base.h, in the definitions for spooling conversions and streaming
  formats we do not supply unnecessary parameters (output data
  formats) to the filter functions any more, but NULL
  instead. Currently, these filter functions ate cfFilterPDFToPDF()
  and cfFilterPDFToRaster().

- To assure that the filter functions get the information about input
  and output data types, we populate content_type and
  final_content_type of the filter data, in both spooling
  (pr_filter()) and streaming (pr_rpreparejob()) printing mode.

- We do not call ppdFilterLoadPPD() in pr_create_job_data() any more
  but now only after we have populated content_type and
  final_content_type in the pr_filter() and pr_rpreparejob(), the
  inout and output format information is needed by ppdFilterLoadPPD().

- In pr_filter() we log "Printing job in spooling mode" right before
  the pr_create_job_data() call now, as it is done at the same place for
  streaming mode in pr_rpreparejob().

- In streaming mode (pr_rpreparejob()) we log both the data format
  which we create from the raster input to feed into the filter chain
  and the format coming out of it in the end.

In addition, a double-free crasher bug got fixed. We freed the name of
the temporary copy of the printer's PPD file at the end of the job,
while the life of the temporary copy is for the whole session of the
Printer Application (or at least for the life of the print queue). So
a filter or backend needing this copy made the Printer Application
crash on the second job.