Temporarily de-activate supply read-out support, done missing part
Supply-level read-out is not yet completed and can make the Printer
Application hang on some printers and cause 100% CPU usage after
printing a job.

On the temporary de-activation of supply-level read-out a part was
forgotten (pr_update_status()) and this causes 100% CPU usage (in the
Avahi sub-thread) when a CUPS backend (PR_COPTIONS_CUPS_BACKENDS) and
not a PAPPL backend is used.

Now the supply-level read-out functionality is also de-activated in
pr_update_status() and this fixes the 100%-CPU problem.

Note that also a PAPPL bug of not resetting the timer after cleaning
up the job history, leading to permanent cleaning action, caused 100%
CPU usage (https://github.com/michaelrsweet/pappl/issues/218). So make
sure PAPPL is also up-to-date.