Update for text-type vendor options fix and handling installable accessories
Before, there was a bug that the web interface did not display
text-style vendor options, like user codes, fax numbers, ...:

https://github.com/michaelrsweet/pappl/issues/142

I abused it also for saving the configuration of installable
accessories ("Installable Options" group in the PPD files) on
PostScript printers without making these PPD options visible on the
"Printing Defaults" web interface page.

I also did a fix for this bug in a way that both the text options get
visisble in the web interface and one can hide away extra config data
in the state file. I added the patch to the mentioned issue report and
also applied it to all the Snaps of retro-fitting Printer Applications
in the Snap Store.

Now Michael Sweet actually fixed the bug, but not using the
patch. With Michael's fix the text options get visible, but one cannot
hide extra data in the state file.

But this is not a problem, as there is the papplPrinterOpenFile() API
function in PAPPL to create extra per-printer configuration files to
accommodate any printer/driver configuration settings which are not
related to anything mappable to IPP attributes.

With this commit it is made use of papplPrinterOpenFile() to save the
installable accessory configuration in separate files. Now the patches
on the Snaps are not needed any more.

Also the state dir (Usually /var/lib/<PRINTER_APP_NAME>/) can be set
via STATE_DIR environment variable or "status-directory" command line
option, to be able to let these extra config files go to the right
place.

Thanks, Michael Sweet, for your hint in this issue report.