Renamed all API functions and constants
To make sure that the resources of libpappl-retrofit do not conflict
with the ones of any other library used by a CUPS-driver-retro-fitting
Printer Application all functions and constants of libpappl-retrofit
got renamed to be unique to libpappl-retrofit. The data types already
have appropriate names.

Here we follow the rules of CUPS and cups-filters (to get unique rules
for all libraries by OpenPrinting, see file DEVELOPING.md in CUPS
source repository): API functions are in camelCase and with "pr"
prefix, data types all-lowercase, with '_' as word separator, and
constants are all-uppercase, also with '_' as word separator, and with
"PR_" prefix. Library-internal (private) functions/constants have an
additionsl underscore "_" at the beginning of the name ("_pr...").