Use PAPPL's papplLocGetDefaultMediaSizeName() for default page size
PAPPL has an SAPI function to determine the default paper size
depending on the user's location now, and this function is even more
accurate in knowing which countries use Letter and not A4.

Therefore we have switched to PAPPL's function now and removed our
library-internal _prDefaultPaperSize() function now.

Also fixed the check for the presence of the "PageSize" option in the
PPD file.