README.md: Updates for CUPS 3.x and fixes
To prepare the 1.0b1 release of pappl-retrofit README.md got updated
and fixed now.

Especially CUPS 3.x is now mentioned as one of the environments which
does not support PPD files and classic drivers any more. There are
also links to the repositories of the components of CUPS 3.x now.

Also the fact that cups-filters 2.x got split up into several
repositories is taken into account now. Recommended version of
libcupsfilters and libppd is 2.0b1 for now.

Otherwise the already existing Printer Applications using
pappl-retrofit are now mentioned as programming examples in their own
section, not in the middle of the section about the Legacy Printer
Application.

Many typos and the line breaks got fixed.