Support for Legacy Printer Application as system daemon
For many users the Legacy Printer Application will be part of their
day-to-day printing as they use it to make a classically installed
(proprietary) CUPS driver available to CUPS 3.x or to the CUPS Snap
with it. These users need it as an auto-starting, permanently running
system daemon.

This commit adds a systemd *.service file and an install mode for
installing the Legacy Printer Application as system daemon. The mode
is triggered by the --enable-legacy-printer-app-as-daemon option for
./configure. Then the *.service file gets installed and the executable
legacy-printer-app gets installed into /usr/(local/)sbin/.