Added support for libcups3 (libcups of CUPS 3.x)  (#15)
With these changes pappl-retrofit can be built either with libcups2
(libcups of CUPS 2.x) or libcups3 (libcups of CUPS 3.x).

    NOTE: Any further development on the code has to be done using the
    API of libcups3, and for libcups2 support any additional
    conversion has to get defined in cupsfilters/libcups2-private.h
    (if C code is needed for the conversion, it goes into
    cupsfilters/libcups2.c, which has to be added in that case).

The actual code is now fully converted to the API of libcups3. To
allow building it with libcups2 and to avoid having it full of "#ifdef
HAVE_LIBCUPS2" conditionals, we have added the header file
libcups2-private.h file, which gets only applied (by "#ifdef
HAVE_LIBCUPS2") when we build with libcups2.

It converts the code to the libcups2 API by

- Applying "#define ..." macro definitions for all renamings of
  functions, data types, and constants

- Converting the "cupsCreateTempFd(NULL, NULL, buffer, bufsize)" calls
  to "cupsTempFd(buffer, bufsize)" calls (new function has more
  arguments).

- Introducing the type "cups_len_t", as "int" for libcups2 and
  "size_t" for libcups3. In some functions there are parameters which
  make a hard switch from "int" to "size_t" with the transition to
  libcups3. For those we use this type.

libcups3 abondoned the back- and sidechannel functionality, as this
was specific to classic CUPS filters. Therefore we have copied the files
cups/sidechannel.h, cups/sidechannel.c, and cups/backchannel.c
from the CUPS 2.4.x GIT branch into the new files
pappl-retorfit/cups-side-back-channel.c and
pappl-retrofit/cups-side-back-channel-private.h. in these file we have
renaming all functions, data types, and constants to match
pappl-retrofit and replaced a CUPS-internal buffer facility by plain
malloc() and free(). So now it builds without warnings when using
libcups3 and should as well build with libcups2 as there are no name
clashes due to the renaming.

Also some general compiler warnings got taken care, of, a "free()"
before the last access in "_prCUPSDevList()", an uninitialized variable
in "_prPrinterWebDeviceConfig()", and too small buffer for "snprintf()"
in "_prSystemWebAddPPD()".

In configure.ac we check for libcups3 via pkg-config first and if not
present for libcups2 via cups-config. In both cases we set all relevant
variables, as compiler/linker arguments, directories, ... and in case
of libcups2 we define HAVE_LIBCUPS2 in config.h. Also checking for
the "poll()" function now, needed for the Back- and side channel code
which got moved into pappl-retrofit.